add diagnostics, rem `proto` module
