--- api/callback/src/proto/mod.rs
@@ -1,2 +1,0 @@
-pub mod into;
-pub mod proxy;
